refactor(reporting): move health to views/reporting/health.py
Split api_sources_health, api_source_schema, _probe_engine, and
_source_used_tables out of nx_lib/views/reporting/__init__.py into
health.py, part of beautify-phase-2a Task 3. __init__.py delegates
route registration to health.register_routes(app).

Caught a real regression before it landed: dropping mapping_config,
db_schema, and time as "unused" also dropped engine_statistics_db,
which nx_lib/reporting/runner.py still reads off
nx_lib.views.reporting at call time (rv.engine_statistics_db) --
caught by tests/unit/test_reporting_shared_exports.py's explicit
attribute-resolution contract test, restored the import, and added it
to __all__. No test patches any of the moved names by string.
