+++ nx_lib/views/reporting/__init__.py
@@ -32,8 +32,6 @@ live-SQL sandbox helpers, ``_get_effective_source``) move to a shared module
 then.
 """
 
-import time
-
 from flask import (
     current_app,
     jsonify,
@@ -42,16 +40,14 @@ from flask import (
 )
 from flask_babel import gettext as _
 
-from ... import mapping_config
 from ...db import engine_statistics_db
 from ...extensions import limiter
 from ...i18n import get_locale
-from ...reporting import db_schema
 from ...reporting.catalog import fetch_docprocessing_catalog
 from ...reporting.sources import accessible
 from ...reporting.table_query import TableQueryError, build_distinct_query, table_source_catalog
 from ...security import has_permission, require_permission
-from . import admin_registry, ai, export, pages, reports, run, schedules
+from . import admin_registry, ai, export, health, pages, reports, run, schedules
 from ._shared import (
     _CURATED_ENGINES,
     _METRIC_LABEL_ATTRS,
@@ -126,6 +122,7 @@ __all__ = [
     "_effective_scope",
     "_effective_sources",
     "_execute",
+    "engine_statistics_db",
     "_extract_agent_artifacts",
     "_forecast_for",
     "_get_effective_source",
@@ -235,119 +232,6 @@ def _accessible_curated_sources():
             }
         )
     return out
-
-
-# ---- Source health (Console sources rail) ---------------------------------
-
-
-def _probe_engine(engine):
-    """(ok, latency_ms, db_name) for one probe round-trip; DB_NAME() rides
-    along because the engines are built from odbc_connect strings whose
-    SQLAlchemy URL carries no database attribute."""
-    if engine is None:
-        return False, None, None
-    t0 = time.perf_counter()
-    try:
-        conn = engine.raw_connection()
-        try:
-            cur = conn.cursor()
-            cur.execute("SELECT DB_NAME()")
-            row = cur.fetchone()
-        finally:
-            conn.close()
-        return True, (time.perf_counter() - t0) * 1000.0, (row[0] if row else None)
-    except Exception:
-        return False, None, None
-
-
-@require_permission("reporting.view")
-def api_sources_health():
-    """Live status dot + latency per accessible source (Console rail).
-    One SELECT-1 probe per distinct engine, shared across sources."""
-    perms = set(session.get("permissions", []))
-    sources = accessible(_effective_sources(), perms)
-    probes = {}  # id(engine) -> (ok, ms)
-
-    def probe(engine):
-        key = id(engine)
-        if key not in probes:
-            probes[key] = _probe_engine(engine)
-        return probes[key]
-
-    out = []
-    for s in sources:
-        if s["kind"] == "sql":
-            engine = _SQL_TARGET_ENGINES.get(s.get("target", "statistics"))
-        else:
-            engine = _CURATED_ENGINES.get(s.get("engine"), engine_statistics_db)
-        ok, ms, db_name = probe(engine)
-        out.append(
-            {
-                "id": s["id"],
-                "ok": ok,
-                "latencyMs": round(ms, 1) if ms is not None else None,
-                "db": db_name,
-            }
-        )
-    return jsonify({"sources": out})
-
-
-def _source_used_tables(src):
-    """Qualified table names the reporting layer actually reads for `src`.
-
-    Two registries know this: the source's own `BaseObject` (the generic
-    `table` provider's target) and `dbo.ProcessSources.TableName` (the
-    statistik table per client/process, migration 0074). Names from a
-    different database simply won't match anything when the payload is
-    filtered, so all of them can be thrown in together.
-    """
-    names = set()
-    if src.get("baseObject"):
-        names.add(src["baseObject"])
-    reg = mapping_config.registry()
-    if reg is not None:
-        names |= {ps.table for ps in reg.sources.values() if ps.table}
-    return names
-
-
-@require_permission("reporting.sources.schema")
-def api_source_schema(source_id):
-    """Tables, columns and foreign keys of the database behind one source.
-
-    Feeds the Console's source visualizer (click a rail card): a filterable
-    table list and an ER diagram. Read-only catalog queries on the same engine
-    the source itself reads from, so no new credential surface -- but the
-    schema of a whole database is more than the source's own fields, hence its
-    own grant on top of the source's permission.
-    """
-    perms = set(session.get("permissions", []))
-    src = next(
-        (s for s in accessible(_effective_sources(), perms) if s.get("id") == source_id),
-        None,
-    )
-    if src is None:
-        return jsonify({"error": _("Not authorized for this source")}), 403
-    if src["kind"] == "sql":
-        engine = _SQL_TARGET_ENGINES.get(src.get("target", "statistics"))
-    else:
-        engine = _CURATED_ENGINES.get(src.get("engine"), engine_statistics_db)
-    if engine is None:
-        return jsonify({"error": _("This source's database is not configured.")}), 503
-    try:
-        conn = engine.raw_connection()
-    except Exception as e:
-        current_app.logger.warning(f"reporting schema: connect failed for {source_id}: {e}")
-        return jsonify({"error": _("Could not reach this database.")}), 503
-    try:
-        payload = db_schema.filter_used(db_schema.introspect(conn), _source_used_tables(src))
-    except Exception as e:
-        current_app.logger.error(f"reporting schema: introspection failed for {source_id}: {e}")
-        return jsonify({"error": _("Could not read this database's schema.")}), 502
-    finally:
-        conn.close()
-    payload["source"] = source_id
-    payload["label"] = src.get("label")
-    return jsonify(payload)
 
 
 def _labeled_field_values(rows, allowed=None):
@@ -465,13 +349,4 @@ def register_routes(app):
     export.register_routes(app)
     reports.register_routes(app)
     schedules.register_routes(app)
-    app.add_url_rule(
-        "/api/reporting/sources/health",
-        endpoint="reporting_sources_health",
-        view_func=api_sources_health,
-    )
-    app.add_url_rule(
-        "/api/reporting/sources/<source_id>/schema",
-        endpoint="reporting_source_schema",
-        view_func=api_source_schema,
-    )
+    health.register_routes(app)
